14.01 second push

+++ src/main/java/javacode/wallet/dto/WalletMapper.java
@@ -3,6 +3,7 @@ package javacode.wallet.dto;
 import javacode.wallet.model.Wallet;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Component
@@ -11,21 +12,21 @@ public class WalletMapper {
     public static Wallet toWalletModel(WalletCreateDto dto) {
         return Wallet.builder()
                 .walletId(UUID.fromString(dto.getWalletId()))
-                .amount(dto.getAmount())
+                .amount(BigDecimal.valueOf(dto.getAmount()))
                 .build();
     }
 
     public static Wallet toWalletModel(WalletUpdateDto dto) {
         return Wallet.builder()
                 .walletId(UUID.fromString(dto.getWalletId()))
-                .amount(dto.getAmount())
+                .amount(BigDecimal.valueOf(dto.getAmount()))
                 .build();
     }
 
     public static WalletResponseDto toWalletResponseDto(Wallet wallet) {
         return WalletResponseDto.builder()
                 .walletId(String.valueOf(wallet.getWalletId()))
-                .amount(wallet.getAmount())
+                .amount(wallet.getAmount().longValue())
                 .build();
 
     }
