+++ src/main/java/javacode/wallet/dto/WalletResponseDto.java
@@ -4,14 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Builder
 @Data
 @AllArgsConstructor
 public class WalletResponseDto {
 
     private String walletId;
-    private BigDecimal amount;
+    private Long amount;
 
 }
